requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
blist==1.3.6
boto==2.32.1
bz2file==0.98
cassandra-driver==2.5.0
certifi==2017.4.17
configparser==3.5.0
cqlengine==0.21.0
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
docopt==0.6.2
EasyProcess==0.2.3
entrypoints==0.2.2
enum34==1.1.6
fasttext==0.8.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==2.2.0
gensim==2.2.0
h5py==2.6.0
html5lib==0.999999999
inspect-it==0.3.2
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.0
lmdb==0.92
Mako==1.0.6
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
mxnet-cu80==0.10.0
MySQL-python==1.2.5
nbconvert==5.2.1
nbformat==4.3.0
nltk==3.0.1
nose==1.3.7
notebook==5.0.0
numpy==1.13.0
opencv-python==3.2.0.7
pandas==0.20.1
pandocfilters==1.4.1
pathlib2==2.2.1
patool==1.12
pbr==3.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.5.3
prettytable==0.7.2
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
Pygments==2.2.0
Pympler==0.5
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyunpack==0.1.2
PyYAML==3.11
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.6.0
scandir==1.5
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.3
SQLAlchemy==0.9.7
statsd==3.0.1
subprocess32==3.2.7
tensorflow==1.1.0
tensorflow-gpu==1.1.0
termcolor==1.0.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wget==3.2
widgetsnbextension==2.0.0
